# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "Message": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "Message",
        "type": "object"
      },
...
      "ModelCatalog": {
        "additionalProperties": false,
        "properties": {
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ModelProfile"
            },
            "title": "Profiles",
            "type": "array"
          }
        },
        "required": [
          "profiles"
        ],
        "title": "ModelCatalog",
        "type": "object"
      },
      "ModelProfile": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,62}$",
            "title": "Id",
            "type": "string"
          },
          "max_output_tokens": {
            "default": 2048,
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "model_ref": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 3,
                "pattern": "^[^/\\s]+/.+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Ref"
          },
          "task_type": {
            "enum": [
              "actor",
              "game_master",
              "wiki_maintenance",
              "editor",
              "writer"
            ],
            "title": "Task Type",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "timeout_seconds": {
            "default": 60,
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "task_type"
        ],
        "title": "ModelProfile",
        "type": "object"
      },
      "ModelRequest": {
        "additionalProperties": false,
        "properties": {
          "max_output_tokens": {
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "maxItems": 128,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "output_schema": {
            "anyOf": [
              {
                "maxLength": 131072,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Schema"
          },
          "profile_id": {
            "minLength": 1,
            "title": "Profile Id",
            "type": "string"
          },
          "task_type": {
            "enum": [
              "actor",
              "game_master",
              "wiki_maintenance",
              "editor",
              "writer"
            ],
            "title": "Task Type",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "timeout_seconds": {
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "integer"
          }
        },
        "required": [
          "profile_id",
          "task_type",
          "messages",
          "max_output_tokens",
          "timeout_seconds"
        ],
        "title": "ModelRequest",
        "type": "object"
      },
      "ModelResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "model_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Ref"
          },
          "parsed_output": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/ModelUsage"
          }
        },
        "required": [
          "profile_id",
          "content"
        ],
        "title": "ModelResponse",
        "type": "object"
      },
      "ModelUsage": {
        "additionalProperties": false,
        "properties": {
          "completion_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Completion Tokens",
            "type": "integer"
          },
          "prompt_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "total_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "title": "ModelUsage",
        "type": "object"
      },
...
      "ProjectModelPolicy": {
        "additionalProperties": false,
        "properties": {
          "agent_profile_ids": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Agent Profile Ids",
            "type": "object"
          },
          "schema_version": {
            "default": 1,
            "maximum": 1.0,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "task_profile_ids": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "enum": [
                "actor",
                "game_master",
                "wiki_maintenance",
                "editor",
                "writer"
              ]
            },
            "title": "Task Profile Ids",
            "type": "object"
          }
        },
        "title": "ProjectModelPolicy",
        "type": "object"
      },
...
      "UsageTotals": {
        "additionalProperties": false,
        "properties": {
          "completion_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Completion Tokens",
            "type": "integer"
          },
          "prompt_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "requests": {
            "default": 0,
            "minimum": 0.0,
            "title": "Requests",
            "type": "integer"
          },
          "total_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "title": "UsageTotals",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local story-domain sidecar API",
    "title": "AI Story Evolution Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/models/catalog": {
      "get": {
        "operationId": "get_model_catalog_models_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalog"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Model Catalog",
        "tags": [
          "models"
        ]
      }
    },
    "/models/complete": {
      "post": {
        "operationId": "complete_model_models_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Complete Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/profiles": {
      "get": {
        "operationId": "get_model_profiles_models_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelProfile"
                  },
                  "title": "Response Get Model Profiles Models Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Model Profiles",
        "tags": [
          "models"
        ]
      }
    },
    "/models/profiles/{profile_id}": {
      "put": {
        "operationId": "put_model_profile_models_profiles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Put Model Profile",
        "tags": [
          "models"
        ]
      }
    },
    "/models/stream": {
      "post": {
        "operationId": "stream_model_models_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Stream Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/usage": {
      "get": {
        "operationId": "model_usage_models_usage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageTotals"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Model Usage",
        "tags": [
          "models"
        ]
      }
    },
...
    "/projects/{project_id}/model-policy": {
      "get": {
        "operationId": "get_project_model_policy_projects__project_id__model_policy_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModelPolicy"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Project Model Policy",
        "tags": [
          "models"
        ]
      },
      "put": {
        "operationId": "put_project_model_policy_projects__project_id__model_policy_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectModelPolicy"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModelPolicy"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Put Project Model Policy",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}